database/requirements.txt
# iBrood Database API Requirements
# For Render PostgreSQL deployment

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
asyncpg>=0.29.0
pydantic[email]>=2.5.3
python-dotenv>=1.0.0
